refactor(tree-grid): Rename barrel file
