+++ projects/igniteui-angular/src/lib/data-operations/data-util.ts
@@ -11,7 +11,7 @@ import { IGroupByRecord } from './groupby-record.interface';
 import { IGroupingState } from './groupby-state.interface';
 import { ISortingExpression } from './sorting-expression.interface';
 import { FilteringStrategy } from './filtering-strategy';
-import { ITreeGridRecord } from '../grids/tree-grid';
+import { ITreeGridRecord } from '../grids/tree-grid/public_api';
 import { cloneValue, mergeObjects } from '../core/utils';
 import { Transaction, TransactionType, HierarchicalTransaction } from '../services/transaction/transaction';
 import { getHierarchy, isHierarchyMatch } from './operations';
