+++ projects/igniteui-angular/src/public_api.ts
@@ -70,7 +70,7 @@ export * from './lib/grids/hiding/hiding.module';
 export * from './lib/grids/common/enums';
 export * from './lib/grids/common/events';
 export * from './lib/grids/grid/public_api';
-export * from './lib/grids/tree-grid/index';
+export * from './lib/grids/tree-grid/public_api';
 export * from './lib/grids/hierarchical-grid/public_api';
 export * from './lib/grids/columns/templates.directive';
 export * from './lib/grids/columns/column.component';
